# frontend/public/swagger.yml
openapi: 3.0.3
info:
  title: Breezy API
  description: API REST du réseau social Breezy
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Serveur local de développement
...
paths:
...
  /posts/{id}:
    delete:
      tags: [Posts]
      summary: Supprimer un post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Supprimé
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...